Use local python and npm if docker is not available

+++ brownie.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
 
-docker run --rm --platform linux/amd64 -v $PWD:/sismondi ghcr.io/c4dt/brownie:latest "$@"
+if docker ps &> /dev/null; then
+  docker run --rm --platform linux/amd64 -v $PWD:/sismondi ghcr.io/c4dt/brownie:latest "$@"
+else
+  brownie "$@"
+fi
